% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentation_functions.R
\name{augment_tree}
\alias{augment_tree}
\title{augment a tree with additional branches}
\usage{
augment_tree(brts, pars, model, soc)
}
\arguments{
\item{brts}{branching times of original tree}

\item{pars}{parameters}

\item{model}{chosen model [rpd1, rpd5, rpd5c]}

\item{soc}{number of species at root (??)}
}
\value{
new tree
}
\description{
augment a tree with additional branches
}
\keyword{augment}
\keyword{tree}
