% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentation_functions.R
\name{augment_tree_tj}
\alias{augment_tree_tj}
\title{augment a tree with additional branches, improved version by TJ}
\usage{
augment_tree_tj(brts, pars, model, soc)
}
\arguments{
\item{brts}{branching times of original tree}

\item{pars}{parameters}

\item{model}{chosen model [rpd1, rpd5, rpd5c]}

\item{soc}{number of species at root (???)}
}
\value{
new tree
}
\description{
augment a tree with additional branches, improved version by TJ
}
\keyword{augment}
\keyword{tree}
