% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMPHASIS_package.R
\name{emphasis}
\alias{emphasis}
\title{perform an emphasis analysis}
\usage{
emphasis(input, file = ".RData", print_process = TRUE, n_it = NULL, tol = 0.01)
}
\arguments{
\item{input}{input object}

\item{file}{input file}

\item{print_process}{verbose output?}

\item{n_it}{number of iterations}

\item{tol}{tolerance}
}
\value{
parameters
}
\description{
perform an emphasis analysis
}
