% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_importance_sampler.R
\name{sim.branchingtimes.and.events}
\alias{sim.branchingtimes.and.events}
\title{simulate branching times and events}
\usage{
sim.branchingtimes.and.events(S = S, ct, p)
}
\arguments{
\item{S}{number of psecies}

\item{ct}{unknown}

\item{p}{probability}
}
\value{
tree data frame
}
\description{
simulate branching times and events
}
\keyword{export}
